conftest for api folder to share setup with other tests
